Adapted to pedantic C compilation
Former-commit-id: cdd397d73cb7d7ff6ddfd9dc5e2c5146c138f0ee

+++ external/sqliteExt/extension-functions.c
@@ -106,7 +106,7 @@ Original code 2006 June 05 by relicoder.
 
 */
 
-//#include "config.h"
+/* #include "config.h" */
 
 #define COMPILE_SQLITE_EXTENSIONS_AS_LOADABLE_MODULE 1
 #define HAVE_ACOSH 1
@@ -1942,6 +1942,6 @@ int double_cmp(const void *a, const void *b){
 
 void print_elem(void *e, int64_t c, void* p){
   int ee = *(int*)(e);
-  printf("%d => %lld\n", ee,c);
+  printf("%d => %lld\n", ee,(long long int)c);
 }
